requirements.txt
kafka-python
python-binance
numpy